`sappho-ast`: split out common expr types into individual mods.
